(174) Begin file_cache refactor, create file_controller sub-module
