+++ src/lib.rs
@@ -43,7 +43,7 @@ pub use build::BuildArgs;
 
 mod build;
 pub mod config;
-mod file_cache;
+mod file_controller;
 mod listings;
 mod highlight;
 mod server;
